Added person removal functionality
Not tested
