+++ PART/Source/CprBroker/CprBroker.sln
@@ -88,8 +88,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tracking.Tests", "Tracking.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DBR.Tests.Comparison", "DBR.Tests.Comparison\DBR.Tests.Comparison.csproj", "{8155A4BE-AC2D-4950-A930-E0062655B60C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Slet", "Slet\Slet.csproj", "{900BF373-C92D-4E5B-B72C-63E0815B67E4}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug.Mono|.NET = Debug.Mono|.NET
@@ -875,42 +873,6 @@ Global
 		{8155A4BE-AC2D-4950-A930-E0062655B60C}.Release|Win32.Build.0 = Release|Any CPU
 		{8155A4BE-AC2D-4950-A930-E0062655B60C}.Release|x86.ActiveCfg = Release|Any CPU
 		{8155A4BE-AC2D-4950-A930-E0062655B60C}.Release|x86.Build.0 = Release|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug.Mono|.NET.ActiveCfg = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug.Mono|.NET.Build.0 = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug.Mono|Any CPU.ActiveCfg = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug.Mono|Any CPU.Build.0 = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug.Mono|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug.Mono|Mixed Platforms.Build.0 = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug.Mono|Mono.ActiveCfg = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug.Mono|Mono.Build.0 = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug.Mono|Win32.ActiveCfg = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug.Mono|Win32.Build.0 = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug.Mono|x86.ActiveCfg = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug.Mono|x86.Build.0 = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug|.NET.ActiveCfg = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug|.NET.Build.0 = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug|Mono.ActiveCfg = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug|Mono.Build.0 = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug|Win32.ActiveCfg = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug|Win32.Build.0 = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Debug|x86.Build.0 = Debug|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Release|.NET.ActiveCfg = Release|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Release|.NET.Build.0 = Release|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Release|Mono.ActiveCfg = Release|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Release|Mono.Build.0 = Release|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Release|Win32.ActiveCfg = Release|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Release|Win32.Build.0 = Release|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Release|x86.ActiveCfg = Release|Any CPU
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -941,7 +903,6 @@ Global
 		{A983F732-5CD5-4DE1-B435-A89E1E18C5D8} = {284ACAE2-9300-447C-BD74-13A27006DA38}
 		{119AC305-96EE-4A08-9FCF-060A826EE82C} = {E1F110C7-3655-43A2-8C95-50FD661D7CE0}
 		{8155A4BE-AC2D-4950-A930-E0062655B60C} = {3389B1C0-EC1E-4298-86E9-27C68859AA4E}
-		{900BF373-C92D-4E5B-B72C-63E0815B67E4} = {3AF1BAC9-AF91-421F-8BF5-3C2D57EC95B8}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {407FEDBD-EEAF-4298-8B64-50C87B7950F4}
