+++ PART/Source/CprBroker/Tracking/Tracking.csproj
@@ -71,7 +71,6 @@
     <Compile Include="CleanupSettings.cs" />
     <Compile Include="Constants.cs" />
     <Compile Include="ITrackingDataProvider.cs" />
-    <Compile Include="PersonRemover.cs" />
     <Compile Include="PersonTrack.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Settings.Designer.cs">
@@ -81,6 +80,8 @@
     </Compile>
     <Compile Include="ReadInstance.cs" />
     <Compile Include="Properties\SettingsUtilities.cs" />
+    <Compile Include="RemovePersonDataProvider.cs" />
+    <Compile Include="RemovePersonDataProvider.Helpers.cs" />
     <Compile Include="RemovePersonFacadeMethodInfo.cs" />
     <Compile Include="RemovePersonItem.cs" />
     <Compile Include="RemovePersonManager.cs" />
@@ -109,9 +110,6 @@
       <Project>{da42d8bb-24d3-4416-a0bd-b98857f249ad}</Project>
       <Name>PartInterface</Name>
     </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="TrackingDataProvider.Helpers.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
