+++ PART/Source/CprBroker/Web/web.config
@@ -64,7 +64,7 @@
         <add type="CprBroker.Engine.Events.DataChangeEventManager, CprBroker.PartInterface" />
         <add type="CprBroker.Engine.Local.ApplicationDataProvider, CprBroker.Engine" />
         <add type="CprBroker.Engine.Ping.PingDataProvider, CprBroker.Engine" />
-        <add type="CprBroker.Providers.Slet.RemovePersonDataProvider, CprBroker.Providers.Slet" />
+        <add type="CprBroker.Slet.RemovePersonDataProvider, CprBroker.PartInterface.Tracking" />
       </knownTypes>
     </dataProviders>
   </dataProvidersGroup>
